Create an API that will fetch total ET and milestone of the specific sprint and Display on burn up chart.
