+++ bff/src/test/controllers/ProjectController/getActiveSprintData.test.ts
@@ -1,339 +1,13 @@
 import axios from 'axios'
-import { Request } from 'express'
 import { DateTime } from 'luxon'
 import httpMocks from 'node-mocks-http'
-import ProjectController from '../../controllers/ProjectController'
-import { CustomTypedResponse } from '../../utils/interfaces'
-import projectTestData from '../constants/projectTestData.json'
+import ProjectController from '../../../controllers/ProjectController'
+import { CustomTypedResponse } from '../../../utils/interfaces'
+import projectTestData from '../../constants/projectTestData.json'
 
 const projectController = new ProjectController()
 
 jest.mock('axios')
-
-describe('When calling getProjectsById() function', () => {
-  let request, response: CustomTypedResponse
-
-  describe('if no user found', () => {
-    beforeAll(async () => {
-      ;(axios as any).mockResolvedValueOnce(
-        // user_id
-        Promise.reject()
-      )
-
-      request = httpMocks.createRequest({
-        method: 'GET',
-        url: '/projects/:id',
-        params: {
-          id: 1
-        },
-        query: {
-          service: 'backlog'
-        }
-      })
-
-      response = httpMocks.createResponse()
-      await projectController.getProjectById(request, response)
-    })
-
-    it('should return status of 403', () => {
-      expect(response.statusCode).toBe(403)
-    })
-
-    it('should respond "Unauthorized access" as validation error', () => {
-      const data = response._getJSONData()
-      expect(data).toHaveProperty('message', 'Unauthorized access')
-    })
-  })
-
-  describe('if no service was provided', () => {
-    beforeAll(async () => {
-      ;(axios as any).mockResolvedValueOnce(
-        // user_id
-        Promise.resolve({ data: { id: 1 } })
-      )
-
-      request = httpMocks.createRequest({
-        method: 'GET',
-        url: '/projects/:id',
-        params: {
-          id: 1
-        }
-      })
-
-      response = httpMocks.createResponse()
-      await projectController.getProjectById(request, response)
-    })
-
-    it('should return status of 400', () => {
-      expect(response.statusCode).toBe(400)
-    })
-
-    it('should respond "Service information not provided." as validation error', () => {
-      const data = response._getJSONData()
-      expect(data).toHaveProperty('message', 'Service information not provided.')
-    })
-  })
-
-  describe('if ID exist in the database', () => {
-    beforeAll(async () => {
-      ;(axios as any)
-        .mockResolvedValueOnce(
-          // user_id
-          Promise.resolve({ data: { id: 1 } })
-        )
-        .mockResolvedValueOnce(Promise.resolve({ data: projectTestData.sampleProject }))
-
-      request = httpMocks.createRequest({
-        method: 'GET',
-        url: '/projects/:id',
-        params: {
-          id: 1
-        },
-        query: {
-          service: 'backlog'
-        }
-      })
-
-      response = httpMocks.createResponse()
-      await projectController.getProjectById(request, response)
-    })
-
-    it('should return status of 200', () => {
-      expect(response.statusCode).toBe(200)
-    })
-
-    it('should return project details', () => {
-      const data = response._getData()
-      expect(JSON.stringify(data)).toBe(JSON.stringify(projectTestData.sampleProject))
-    })
-  })
-
-  describe('if ID does not exist in the database', () => {
-    beforeAll(async () => {
-      ;(axios as any)
-        .mockResolvedValueOnce(
-          // user_id
-          Promise.resolve({ data: { id: 1 } })
-        )
-        .mockResolvedValueOnce(
-          Promise.reject({ response: { data: { message: 'No Data Found' }, status: 404 } })
-        )
-
-      request = httpMocks.createRequest({
-        method: 'GET',
-        url: '/projects/:id',
-        params: {
-          id: '111111'
-        },
-        query: {
-          service: 'backlog'
-        }
-      })
-
-      response = httpMocks.createResponse()
-      await projectController.getProjectById(request, response)
-    })
-
-    it('should return status 404', () => {
-      expect(response.statusCode).toBe(404)
-    })
-
-    it('should respond "No Data Found" as validation error', () => {
-      const data = response._getJSONData()
-      expect(data).toHaveProperty('message', 'No Data Found')
-    })
-  })
-
-  describe('if ID is invalid', () => {
-    beforeAll(async () => {
-      ;(axios as any)
-        .mockResolvedValueOnce(
-          // user_id
-          Promise.resolve({ data: { id: 1 } })
-        )
-        .mockResolvedValueOnce(
-          Promise.reject({ response: { data: { message: 'Invalid ID' }, status: 400 } })
-        )
-
-      request = httpMocks.createRequest({
-        method: 'GET',
-        url: '/projects/:id',
-        params: {
-          id: 'test'
-        },
-        query: {
-          service: 'backlog'
-        }
-      })
-
-      response = httpMocks.createResponse()
-      await projectController.getProjectById(request, response)
-    })
-
-    it('should return status 400', () => {
-      expect(response.statusCode).toBe(400)
-    })
-
-    it('should respond "No Data Found" as validation error', () => {
-      const data = response._getJSONData()
-      expect(data).toHaveProperty('message', 'Invalid ID')
-    })
-  })
-})
-
-describe('When calling deleteProjectById() function', () => {
-  let request: Request, response: CustomTypedResponse
-
-  describe('if no user found', () => {
-    beforeAll(async () => {
-      ;(axios as any).mockResolvedValueOnce(
-        // user_id
-        Promise.reject()
-      )
-
-      request = httpMocks.createRequest({
-        method: 'DELETE',
-        url: '/projects/:id',
-        query: { service: 'backlog' },
-        params: { id: '1' }
-      })
-      response = httpMocks.createResponse()
-
-      await projectController.deleteProjectById(request, response)
-    })
-
-    it('should return status of 403', () => {
-      expect(response.statusCode).toBe(403)
-    })
-
-    it('should return "Unauthorized access" as error message', () => {
-      const data = response._getJSONData()
-      expect(data).toHaveProperty('message', 'Unauthorized access')
-    })
-  })
-
-  describe('if no service was provided', () => {
-    beforeAll(async () => {
-      ;(axios as any).mockResolvedValueOnce(
-        // user_id
-        Promise.resolve({ data: { id: 1 } })
-      )
-
-      request = httpMocks.createRequest({
-        method: 'DELETE',
-        url: '/projects/:id',
-        params: { id: '1' }
-      })
-      response = httpMocks.createResponse()
-
-      await projectController.deleteProjectById(request, response)
-    })
-
-    it('should return status of 400', () => {
-      expect(response.statusCode).toBe(400)
-    })
-
-    it('should return "Service information is not provided." as error message', () => {
-      const data = response._getJSONData()
-      expect(data).toHaveProperty('message', 'Service information is not provided.')
-    })
-  })
-
-  describe('if ID does not exist in the database', () => {
-    beforeAll(async () => {
-      ;(axios as any)
-        .mockResolvedValueOnce(
-          // user_id
-          Promise.resolve({ data: { id: 1 } })
-        )
-        .mockResolvedValueOnce(
-          Promise.reject({ response: { data: { message: 'ID does not exist' }, status: 404 } })
-        )
-
-      request = httpMocks.createRequest({
-        method: 'DELETE',
-        url: '/projects/:id',
-        query: { service: 'backlog' },
-        params: { id: '1' }
-      })
-      response = httpMocks.createResponse()
-
-      await projectController.deleteProjectById(request, response)
-    })
-
-    it('should return status of 404', () => {
-      expect(response.statusCode).toBe(404)
-    })
-
-    it('should return expected error message', () => {
-      const data = response._getJSONData()
-      expect(data).toHaveProperty('message', 'ID does not exist')
-    })
-  })
-
-  describe('if invalid ID, letters are not valid, should be number', () => {
-    beforeAll(async () => {
-      ;(axios as any)
-        .mockResolvedValueOnce(
-          // user_id
-          Promise.resolve({ data: { id: 1 } })
-        )
-        .mockResolvedValueOnce(
-          Promise.reject({ response: { data: { message: 'Invalid ID' }, status: 400 } })
-        )
-
-      request = httpMocks.createRequest({
-        method: 'DELETE',
-        url: '/projects/:id',
-        query: { service: 'backlog' },
-        params: { id: 'test' }
-      })
-      response = httpMocks.createResponse()
-
-      await projectController.deleteProjectById(request, response)
-    })
-
-    it('should return status of 400', () => {
-      expect(response.statusCode).toBe(400)
-    })
-
-    it('should return expected error message', () => {
-      const data = response._getJSONData()
-      expect(data).toHaveProperty('message', 'Invalid ID')
-    })
-  })
-
-  describe('if valid ID', () => {
-    beforeAll(async () => {
-      ;(axios as any)
-        .mockResolvedValueOnce(
-          // user_id
-          Promise.resolve({ data: { id: 1 } })
-        )
-        .mockResolvedValueOnce(Promise.resolve({ data: projectTestData.sampleProject }))
-
-      request = httpMocks.createRequest({
-        method: 'DELETE',
-        url: '/projects/:id',
-        query: { service: 'backlog' },
-        params: { id: 1 }
-      })
-      response = httpMocks.createResponse()
-
-      await projectController.deleteProjectById(request, response)
-    })
-
-    it('should return status of 200', () => {
-      expect(response.statusCode).toBe(200)
-    })
-
-    it('should return deleted project details', () => {
-      const data = response._getData()
-
-      expect(JSON.stringify(data)).toBe(JSON.stringify(projectTestData.sampleProject))
-    })
-  })
-})
 
 describe('When calling getActiveSprintData() function', () => {
   let request, response: CustomTypedResponse
@@ -751,19 +425,29 @@ describe('When calling getActiveSprintData() function', () => {
     const diffTime = Math.abs(date2 - date1)
     const diffDays = Math.floor(diffTime / (1000 * 60 * 60 * 24))
 
-    const timeArray = []
-    const datesArray = []
+    const timeArray = [],
+      datesArray = [],
+      burnUp = []
 
     for (let i = 0; i < diffDays; i++) {
       timeArray.push(0)
       const result = new Date(date1)
       result.setDate(result.getDate() + i + 1)
       datesArray.push(result.toLocaleDateString())
+      burnUp.push(5)
     }
 
     expectedResponse.data.actual.push.apply(expectedResponse.data.actual, timeArray)
     expectedResponse.data.estimated.push.apply(expectedResponse.data.estimated, timeArray)
     expectedResponse.dates.push.apply(expectedResponse.dates, datesArray)
+    expectedResponse.burnUpChartData.completedPoints.push.apply(
+      expectedResponse.burnUpChartData.completedPoints,
+      burnUp
+    )
+    expectedResponse.burnUpChartData.totalPoints.push.apply(
+      expectedResponse.burnUpChartData.totalPoints,
+      burnUp
+    )
 
     beforeAll(async () => {
       ;(axios as any)
