+++ fe/styles/project-detail.module.css
@@ -1,5 +1,6 @@
 .burn-up-chart {
-  padding-top: 63px;
+  margin-top: 4em;
+  width: 47%;
 }
 
 .burn-down-chart-select {
@@ -15,7 +16,7 @@
 
 .brn-down-chart {
   margin-top: 4em;
-  width: 50%;
+  width: 47%;
 }
 
 .velocity-chart-velocity {
